Add files via upload
The Coding Research Repository #  
This repository includes a selection of my own codes, codes I've practiced or been inspired by, and software development-related subjects that I find intriguing or difficult.  

 

## My Codes  

 

I wrote the codes in this directory for a variety of purposes. Every file has comments that describe the function and operation of the code.  

 

## Motivated/Used Codes  

 

These are codes I've encountered and worked with from books, tutorials, and other sources. Every file has comments outlining the functionality of the code as well as any knowledge I've gained from using it.  

 

## Fascinating/Uncertain  
I list subjects or codes that I think are especially fascinating or perplexing in this area. Every file or markdown document has my own research, questions, and explanations.